TTAR-0163 Fix return type of coupon price calculate method

+++ src/main/java/com/ttarum/member/domain/coupon/Coupon.java
@@ -23,7 +23,7 @@ public class Coupon {
 
     private int value;
 
-    public double calculatePrice(int originalPrice) {
+    public int calculatePrice(int originalPrice) {
         return couponStrategy.calculate(originalPrice, value);
     }
 }
